+++ src/CfdiUtils/Elements/CartaPorte20/Ubicacion.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace CfdiUtils\Elements\CartaPorte20;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class Ubicacion extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'cartaporte20:Ubicacion';
+    }
+
+    public function getDomicilio(): Domicilio
+    {
+        return $this->helperGetOrAdd(new Domicilio());
+    }
+
+    public function addDomicilio(array $attributes = []): Domicilio
+    {
+        $subject = $this->getDomicilio();
+        $subject->addAttributes($attributes);
+        return $subject;
+    }
+}